requirements.txt
tensorflow>=2.12.0
opencv-python>=4.8.0
numpy>=1.23.0
Pillow>=9.5.0
matplotlib>=3.7.0
streamlit>=1.28.0
scikit-learn>=1.2.0
